feat: move  bookmarks and search markup to previewView
