+++ src/js/views/bookmarksView.js
@@ -1,4 +1,5 @@
 import View from './View.js';
+import previewView from './previewView.js';
 import icons from 'url:../../img/icons.svg';
 
 class BookmarksView extends View {
@@ -7,29 +8,8 @@ class BookmarksView extends View {
 
   _generateMarkup() {
     return this._data
-      .map((result) => {
-        return this.#generateMarkupPreviews(result);
-      })
+      .map((result) => previewView.render(result, false))
       .join('');
-  }
-
-  #generateMarkupPreviews(result) {
-    const id = window.location.hash.slice(1);
-    return `
-        <li class="preview">
-          <a class="preview__link ${
-            result.id === id ? 'preview__link--active' : ''
-          }" href="#${result.id}">
-            <figure class="preview__fig">
-              <img src="${result.image}" alt="${result.title}" />
-            </figure>
-            <div class="preview__data">
-              <h4 class="preview__title">${result.title}</h4>
-              <p class="preview__publisher">${result.publisher}</p>
-            </div>
-          </a>
-        </li>
-    `;
   }
 }
